raspberry/requirements.txt
chess==1.10.0
pyserial==3.5
